validations

+++ App/Controllers/AccController.php
@@ -39,7 +39,36 @@ class AccController
         $number = $request['number'];
         $code = $request['code'];
 
+        if(strlen($fname) < 4 || strlen($lname) < 4){
+            Message::get()->set('danger', 'Vardas ir pavardė turi būti ilgesni nei 3 simboliai');
+            return App::redirect('create');
+        }
+
         $writer = new FileBase('accounts');
+        $accounts = $writer->showAll();
+
+        foreach($accounts as $account){
+            if($account->code == $code){
+                Message::get()->set('danger', 'Sąskaita tokiu asmens kodu jau egzistuoja');
+                return App::redirect('create');
+            }
+        }
+
+        if(strlen($code) != 11){
+            Message::get()->set('danger', 'Asmens kodą turi sudaryti 11 simbolių');
+            return App::redirect('create');
+        }
+
+        if(substr($code, 0, 1) < 3 || (substr($code, 0, 1) > 6)){
+            Message::get()->set('danger', 'Neteisingas asmens kodas');
+            return App::redirect('create');
+        }
+        if(substr($code, 3, 1) > 1){
+            Message::get()->set('danger', 'Neteisingas asmens kodas');
+            return App::redirect('create');
+        }
+        
+        
         $writer->create((object) [
             'fname' => $fname,
             'lname' => $lname,
@@ -54,8 +83,12 @@ class AccController
 
     public function destroy($id){
         $writer = new FileBase('accounts');
+        $account = $writer->show($id);
+        if($account->balance != 0){
+            Message::get()->set('danger', 'Negalima ištrinti sąskaitos kurioje yra lėšų');
+            return App::redirect('accounts');
+        }
         $writer->delete($id);
-
         Message::get()->set('danger', 'Sąskaita ištrinta');
         return App::redirect('accounts');
     }
@@ -72,6 +105,10 @@ class AccController
 
     public function update($id, $request){
         $add = $request['add'];
+        if($add < 0){
+            Message::get()->set('danger', 'Negalima pridėti neigiamos sumos');
+            return App::redirect('accounts'); 
+        }
         $add = intval($add);    
         $writer = new FileBase('accounts');
         $account = $writer->show($id);
@@ -96,6 +133,11 @@ class AccController
         $withdraw = intval($withdraw);    
         $writer = new FileBase('accounts');
         $account = $writer->show($id);
+
+        if($account->balance < $withdraw){
+            Message::get()->set('danger', 'Negalima nuskaičiuoti daugiau nei esamas likutis');
+            return App::redirect('accounts'); 
+        }
         $account->balance -= $withdraw;
         $writer->withdrawupd($id, $account);
